backend/requirement.txt
Flask
python-dotenv
requests
flask-cors
google-generativeai
Werkzeug